+++ src/main/java/kitchenpos/eatinorder/application/EatInOrderService.java
@@ -1,6 +1,6 @@
 package kitchenpos.eatinorder.application;
 
-import kitchenpos.eatinorder.tobe.domain.*;
+import kitchenpos.eatinorder.tobe.domain.order.*;
 import kitchenpos.eatinorder.tobe.domain.ordertable.OrderTable;
 import kitchenpos.eatinorder.tobe.domain.ordertable.OrderTableRepository;
 import kitchenpos.menus.tobe.domain.menu.Menu;
