+++ src/main/java/kitchenpos/eatinorder/tobe/domain/order/LastEatInrderClearedTable.java
@@ -1,18 +1,15 @@
-package kitchenpos.eatinorder.tobe.domain;
+package kitchenpos.eatinorder.tobe.domain.order;
 
 import kitchenpos.annotation.DomainService;
-import kitchenpos.eatinorder.tobe.domain.ordertable.OrderTableRepository;
 
 import java.util.UUID;
 
 @DomainService
 public class LastEatInrderClearedTable implements ClearedTable {
     OrderRepository orderRepository;
-    OrderTableRepository orderTableRepository;
 
-    public LastEatInrderClearedTable(OrderRepository orderRepository, OrderTableRepository orderTableRepository) {
+    public LastEatInrderClearedTable(OrderRepository orderRepository) {
         this.orderRepository = orderRepository;
-        this.orderTableRepository = orderTableRepository;
     }
 
     @Override
