+++ src/main/java/kitchenpos/eatinorder/tobe/domain/order/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorder.tobe.domain;
+package kitchenpos.eatinorder.tobe.domain.order;
 
 import jakarta.persistence.*;
